+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.4] - 2026-03-06
+## [1.0.5] - 2026-03-06
+
+* [Feature] Convert an image into base64
+
+## [1.0.4] - 2026-03-05
 
 * [Technical] Removed TailwindCSS dependencies and utility classes across all components.
 * [Technical] Implemented a custom CSS architecture for layouts and UI elements.